Move benchmarks to their own files.
