+++ relstorage/cache/tests/test_cache.py
@@ -234,7 +234,6 @@ class StorageCacheTests(unittest.TestCase):
         self.assertEqual(data, {})
 
     def test_after_tpc_finish(self):
-        from relstorage.tests.fakecache import data
         from ZODB.utils import p64
         c = self._makeOne()
         c.tpc_begin()
@@ -917,8 +916,8 @@ class LocalClientTests(unittest.TestCase):
             shutil.rmtree(root_temp_dir)
 
     def test_load_and_save_new_dir(self):
-         # automatically create directories as needed
-         self.test_load_and_save(False)
+        # automatically create directories as needed
+        self.test_load_and_save(False)
 
 from relstorage.options import Options
 
@@ -955,341 +954,6 @@ class MockPoller(object):
         return ((oid, tid) for (oid, tid) in self.changes
                 if tid > after_tid and tid <= last_tid)
 
-def local_benchmark():
-    from relstorage.cache.mapping import SizedLRUMapping
-    from relstorage.cache import LocalClient
-    options = MockOptions()
-    options.cache_local_mb = 100
-    options.cache_local_compression = 'none'
-
-    REPEAT_COUNT = 4
-
-    KEY_GROUP_SIZE = 400
-    DATA_SIZE = 1024
-
-    # With 1000 in a key group, and 1024 bytes of data, we produce
-    # 909100 keys, and 930918400 = 887MB of data, which will overflow
-    # a cache of 500 MB.
-
-    # A group size of 100 produces 9100 keys with 9318400 = 8.8MB of data.
-    # Likewise, group of 200 produces 36380 keys with 35.5MB of data.
-
-    # Group size of 400 produces 145480 keys with 142MB of data.
-
-    # Most of our time is spent in compression, it seems.
-    # In the 8.8mb case, populating all the data with default compression
-    # takes about 2.5-2.8s. Using no compression, it takes 0.38 to 0.42s.
-    # Reading is the same at about 0.2s.
-
-    # Before any changes:
-    # cache_local_mb = 500, datasize = 142, comp=none
-    # epop average 3.304353015982391 stddev 0.1057548559581552
-    # mix  average 2.922693134013874 stddev 0.014240008454610707
-    # pop  average 2.2137693379966854 stddev 0.09458639191519878
-    # read average 1.0852473539998755 stddev 0.023173488388016424
-
-    # cache_local_mb = 100, datasize=142, comp=default
-    # epop average 30.283703678986058 stddev 0.349105696895158
-    # mix  average 32.43547729967395 stddev 0.6131160273617585
-    # pop  average 31.683537834013503 stddev 0.9313916809959417
-    # read average 0.7965960823348723 stddev 0.013812922826548332
-
-    # cache_local_mb = 100, datasize=142, comp=none
-    # epop average 3.8289742503354014 stddev 0.138905518890246
-    # mix  average 6.044395485989905 stddev 0.12402917755863634
-    # pop  average 4.849317686690483 stddev 0.3407186386084065
-    # read average 0.7788464936699407 stddev 0.011301417502572604
-
-    # Following numbers are all with 100/142/none (Python 3.4),
-    # except as noted
-
-    # Tracking popularity, but not aging:
-    # epop average 3.8351666433348632 stddev 0.016045702030828404
-    # mix  average 6.063804395322222 stddev 0.05007505835225963
-    # pop  average 4.915782862672738 stddev 0.20628836098923425
-    # read average 0.8606604933350658 stddev 0.01461748647882393
-
-    # Aging periodically, adjusted for size. We aged three times
-    # during the 'mixed' workload at about  0.024s each. That should be
-    # linear, so the 800MB case would take 0.14s....but it would only be done
-    # every 9,091,000 operations.
-
-    # I noticed differences accounted for by hash ranomization between runs.
-    # From now on, run with 'PYTHONHASHSEED=0 python -O ...'
-    # Still same code
-    # epop average 3.896360943307324 stddev 0.05112068256616049
-    # mix  average 6.08575853901372 stddev 0.0651629903238879
-    # pop  average 4.854507520659051 stddev 0.16709270096300968
-    # read average 0.9192146409768611 stddev 0.010830646982195127
-
-    # Eden generation, unoptimized
-    # epop average 8.394099957639506 stddev 0.1772435870640342
-    # mix  average 5.722020475019235 stddev 0.11354930215079416
-    # pop  average 9.779178152016053 stddev 0.2953541870308067
-    # read average 0.9772441836539656 stddev 0.010378042791130002
-
-    # Full segmented LRU, unoptimized
-    # epop average 9.370241302996874 stddev 0.20596681998331154
-    # mix  average 8.483468734678658 stddev 0.051873515883878445
-    # pop  average 9.935747388653303 stddev 0.10084787068640827
-    # read average 0.42481018069277826 stddev 0.0225555561
-
-    # ^ The incredibly low read times are due to *nothing* actually
-    # getting promoted to the protected generation, so there is
-    # nothing to read. Once we pre-populate the protected segment,
-    # that times goes to 3.8 seconds (python 2.7):
-
-    # read average 3.85280092557 stddev 0.29397446809
-
-    # With that pre-population done, and compression back on, we
-    # still find the 3s slowdown for reads, which is unacceptable:
-
-    # epop average 37.56022281331631 stddev 2.3185901813555603
-    # mix  average 32.643460757661764 stddev 0.4262010697672024
-    # pop  average 34.260926674314156 stddev 0.5609715096223009
-    # read average 3.956128183985129 stddev 0.20373283218514573
-
-    # If we add a function for dedicated access to the LRU item,
-    # our times almost become acceptable:
-    # - mix is faster than original;
-    # - pop is faster than original;
-    # - epop is slightly slower than original;
-    # - read is about twice as slow as the original.
-    # epop average 5.056254096988899 stddev 0.07991501317283245
-    # mix  average 3.79573760201068 stddev 0.17857980670336743
-    # pop  average 2.814852144336328 stddev 0.3669570906504935
-    # read average 1.3201618876773864 stddev 0.008809367575729931
-
-    # ^ Sigh, no, those times were a bug. our dedicated LRU access method
-    # was actually returning the MRU item, which means we usually skipped all
-    # the manipulation of lists. With that fixed, and some minor opts, we're at these times:
-    # - mix is faster; everything else is slower; read is *much* slower.
-    # epop average 7.55050067200015 stddev 0.1184560690858881
-    # mix  average 5.364189800301877 stddev 0.015863679783528113
-    # pop  average 5.598568096330079 stddev 0.31457757182924395
-    # read average 3.0701343226634585 stddev 0.19113773305717596
-
-    # PyPy is *wicked* fast at this, BTW
-    # epop average 1.30581800143 stddev 0.0167331686883
-    # mix  average 0.690402587255 stddev 0.0491290787176
-    # pop  average 0.915206670761 stddev 0.126843920891
-    # read average 0.579447031021 stddev 0.0305893529027
-
-    # ""Inlining"" Ring.add into Ring.move_entry_from_other_ring
-    # got us down to about 2.5s on the read test (by not allocating new
-    # CPersistentRing structures).
-    # Introducing the C function ring_move_to_head_from_foreign gives us these numbers:
-    # epop average 6.472030825718927 stddev 0.06240653685961519
-    # mix  average 5.4434908026984585 stddev 0.07889832553078663
-    # pop  average 4.920183609317367 stddev 0.23233663177104144
-    # read average 1.7981388796276103 stddev 0.0625821728436513
-
-    with open('/dev/urandom', 'rb') as f:
-        random_data = f.read(DATA_SIZE)
-
-    key_groups = []
-    key_groups.append([str(i) for i in range(KEY_GROUP_SIZE)])
-    for i in range(1, KEY_GROUP_SIZE):
-        keys = [str(i) + str(j) for j in range(KEY_GROUP_SIZE)]
-        assert len(set(keys)) == len(keys)
-        key_groups.append(keys)
-
-
-    # Recent PyPy and Python 3.6 preserves iteration order of a dict
-    # to match insertion order. If we use a dict for ALL_DATA, this
-    # gives slightly different results due to the key lengths being
-    # different and so things being ejected at slightly different
-    # times (on PyPy, 8 key groups have *no* matches in read() using a dict,
-    # while that doesn't occur in cPython 2.7/3.4). To
-    # make this all line up the same, we preserve order everywhere by using
-    # a list of tuples (sure enough, that change makes 8 groups go missing)
-    # Alternately, if we sort by the hash of the key, we get the iteration order that
-    # CPython used for a dict, making all groups of keys be found in read(). This
-    # keeps the benchmark consistent
-
-    ALL_DATA = {}
-    for group in key_groups:
-        for key in group:
-            ALL_DATA[key] = random_data
-    assert all(isinstance(k, str) for k in ALL_DATA)
-    ALL_DATA = list(ALL_DATA.items())
-    ALL_DATA.sort(key=lambda x: hash(x[0]))
-    print(len(ALL_DATA), sum((len(v[1]) for v in ALL_DATA))/1024/1024)
-
-
-    def do_times(client_type=LocalClient):
-        client = client_type(options)
-        print("Testing", type(client._bucket0._dict))
-
-        def populate():
-            for k, v in ALL_DATA:
-                client.set(k, v)
-
-
-        def populate_empty():
-            c = LocalClient(options)
-            for k, v in ALL_DATA:
-                c.set(k, v)
-
-        def read():
-            # This is basically the worst-case scenario for a basic
-            # segmented LRU: A repeating sequential scan, where no new
-            # keys are added and all existing keys fit in the two parts of the
-            # cache. Thus, entries just keep bouncing back and forth between
-            # probation and protected. It so happens that this is our slowest
-            # case.
-            miss_count = 0
-            for keys in key_groups:
-                res = client.get_multi(keys)
-                #assert len(res) == len(keys)
-                if not res:
-                    miss_count += 1
-                    continue
-                assert res.popitem()[1] == random_data
-
-            if miss_count:
-                print("Failed to get any keys in %d of %d groups"
-                      % (miss_count, len(key_groups)))
-
-            # import pprint
-            # pprint.pprint(client._bucket0.stats())
-            # print("Probation promotes", client._bucket0._probation.promote_count)
-            # print("Probation demotes", client._bucket0._probation.demote_count)
-            # print("Probation removes", client._bucket0._probation.remove_count)
-
-        def mixed():
-            hot_keys = key_groups[0]
-            i = 0
-            for k, v in ALL_DATA:
-                i += 1
-                client.set(k, v)
-                if i == len(hot_keys):
-                    client.get_multi(hot_keys)
-                    i = 0
-
-        import timeit
-        import statistics
-        try:
-            import sys
-            import cProfile, pstats
-            if '--profile' not in sys.argv:
-                raise ImportError
-        except ImportError:
-            class cProfile(object):
-                class Profile(object):
-                    def enable(self): pass
-                    def disable(self): pass
-            class pstats(object):
-                class Stats(object):
-                    def __init__(self, *args): pass
-                    def sort_stats(self, *args): return self
-                    def print_stats(self, *args): pass
-
-
-        number = REPEAT_COUNT
-        def run_func(func):
-            print("Timing func", func)
-            pop_timer = timeit.Timer(func)
-            pr = cProfile.Profile()
-            pr.enable()
-            pop_times = pop_timer.repeat(number=number)
-            pr.disable()
-            ps = pstats.Stats(pr).sort_stats('cumulative')
-            ps.print_stats(.4)
-
-            return pop_times
-
-        times = {}
-        for name, func in (('pop ', populate),
-                           ('epop', populate_empty),
-                           ('read', read),
-                           ('mix ', mixed),):
-            times[name] = run_func(func)
-
-        for name, time in sorted(times.items()):
-
-            print(name, "average", statistics.mean(time), "stddev", statistics.stdev(time))
-
-
-    do_times()
-
-def save_load_benchmark():
-    from relstorage.cache.mapping import SizedLRUMapping as LocalClientBucket
-    from relstorage.cache import _persistence as _Loader
-    from io import BytesIO
-    import os
-    import itertools
-
-    import sys
-    sys.setrecursionlimit(500000)
-    bucket = LocalClientBucket(500*1024*1024)
-    print("Testing", type(bucket._dict))
-
-
-    size_dists = [100] * 800 + [300] * 500 + [1024] * 300 + [2048] * 200 + [4096] * 150
-
-    with open('/dev/urandom', 'rb') as rnd:
-        data = [rnd.read(x) for x in size_dists]
-    data_iter = itertools.cycle(data)
-
-    for j, datum in enumerate(data_iter):
-        if len(datum) > bucket.limit or bucket.size + len(datum) > bucket.limit:
-            break
-        # To ensure the pickle memo cache doesn't just write out "use object X",
-        # but distinct copies of the strings, we need to copy them
-        bucket[str(j)] = datum[:-1] + b'x'
-        assert bucket[str(j)] is not datum
-        #print("Len", len(bucket), "size", bucket.size, "dlen", len(datum))
-
-    print("Len", len(bucket), "size", bucket.size)
-    number = 1
-    import timeit
-    import statistics
-    import cProfile
-    import pstats
-
-    cache_pfx = "pfx"
-    cache_options = MockOptions()
-    cache_options.cache_local_dir = '/tmp'
-    cache_options.cache_local_dir_compress = False
-
-    def write():
-        fname = _Loader.save_local_cache(cache_options, cache_pfx, bucket)
-        os.remove(fname)
-
-
-    def load():
-        b2 = LocalClientBucket(bucket.limit)
-        _Loader.load_local_cache(cache_options, cache_pfx, b2)
-
-    #write_timer = timeit.Timer(write)
-    #write_times = write_timer.repeat(number=number)
-    #print("write average", statistics.mean(write_times), "stddev", statistics.stdev(write_times))
-
-    #read_timer = timeit.Timer(load)
-    #read_times = read_timer.repeat(number=number)
-    #print("read average", statistics.mean(read_times), "stddev", statistics.stdev(read_times))
-
-    #pr = cProfile.Profile()
-    #pr.enable()
-
-    fname = _Loader.save_local_cache(cache_options, cache_pfx, bucket)
-    print("Saved to", fname)
-    #pr.disable()
-    #ps = pstats.Stats(pr).sort_stats('cumulative')
-    #ps.print_stats()
-    #return
-
-    pr = cProfile.Profile()
-    pr.enable()
-    _Loader.load_local_cache(cache_options, cache_pfx, LocalClientBucket(bucket.limit))
-    pr.disable()
-    ps = pstats.Stats(pr).sort_stats('cumulative')
-    ps.print_stats(.4)
-
-
-    #os.remove(fname)
-
 def test_suite():
     suite = unittest.TestSuite()
     suite.addTest(unittest.makeSuite(StorageCacheTests))
@@ -1298,12 +962,4 @@ def test_suite():
     return suite
 
 if __name__ == '__main__':
-    import sys
-    if '--localbench' in sys.argv:
-        local_benchmark()
-    elif '--iobench' in sys.argv:
-        import logging
-        logging.basicConfig(level=logging.DEBUG)
-        save_load_benchmark()
-    else:
-        unittest.main(defaultTest='test_suite')
+    unittest.main(defaultTest='test_suite')
